Add POST /authors endpoint

+++ src/main/resources/db/migration/V1__Create_authors_and_books_tables.sql
@@ -2,7 +2,7 @@ CREATE TABLE authors (
     id SERIAL PRIMARY KEY,
     first_name VARCHAR(255) NOT NULL,
     last_name VARCHAR(255) NOT NULL,
-    age NUMERIC NOT NULL,
+    age INTEGER NOT NULL,
     created TIMESTAMP,
     updated TIMESTAMP
 );
